requirements.txt
streamlit
pymupdf
google.generativeai
regex
youtube-transcript-api
ebooklib
beautifulsoup4